Renamed ViewHeader index file to component name
