+++ src/components/index.tsx
@@ -1,3 +1,3 @@
 export { Link } from './Link/Link'
-export { ViewHeader } from './ViewHeader'
+export { ViewHeader } from './ViewHeader/ViewHeader'
 export { AppHeader } from './AppHeader'
